Tab to spaces. Cast to float in IntVector2::Length() to avoid warning spam.

+++ Source/Urho3D/Math/Vector2.h
@@ -330,7 +330,7 @@ public:
     unsigned ToHash() const { return (unsigned)x_ * 31 + (unsigned)y_; }
 
     /// Return length.
-    float Length() const { return sqrtf(x_ * x_ + y_ * y_); }
+    float Length() const { return sqrtf((float)(x_ * x_ + y_ * y_)); }
 
     /// X coordinate.
     int x_;
